📝 docs(LICENSE): add MIT License to the project
The MIT License is added to the project to grant permission for anyone to use, copy, modify, merge, publish, distribute, sublicense, and/or sell the software. This license ensures that the software is provided "as is" without warranty and the authors or copyright holders are not liable for any claim, damages, or other liability.